Former-commit-id: 91dc648418de4e5207e24f4292c828abf4b46033

+++ src/hicks/td/ui/DisplayInfo.java
@@ -9,8 +9,8 @@ public class DisplayInfo
 
     public static double getScalingFactor()
     {
-        double scaleWidth = 1024 / displayWidth;
-        double scaleHeight = 800 / displayHeight;
+        double scaleWidth = (double) 1024 / displayWidth;
+        double scaleHeight = (double) 800 / displayHeight;
 
         String scaleBy = scaleWidth < scaleHeight ? "width" : "height";
